+++ protocol-designer/src/pages/Designer/ProtocolSteps/StepForm/StepTools/MoveLiquidTools/SecondStepsMoveLiquidTools.tsx
@@ -19,7 +19,7 @@ import {
   getMaxPushOutVolume,
   getMinXYDimension,
   NONE_LIQUID_CLASS_NAME,
-  WATER_LIQUID_CLASS_NAME_V2,
+  WATER_LIQUID_CLASS_NAME,
 } from '@opentrons/shared-data'
 import {
   getPipetteWithTipMaxVol,
@@ -106,7 +106,7 @@ export const SecondStepsMoveLiquidTools = ({
     [
       formData.liquidClass !== NONE_LIQUID_CLASS_NAME
         ? formData.liquidClass
-        : WATER_LIQUID_CLASS_NAME_V2
+        : WATER_LIQUID_CLASS_NAME
     ].byPipette.find(
       ({ pipetteModel }) => pipetteModel === getFlexNameConversion(pipetteSpecs)
     )