+++ protocol-designer/typings/global.d.ts
@@ -18,8 +18,10 @@ declare module '*.md' {
 declare const _FF_ENV_VARS_: Record<string, string | undefined>
 declare const _NODE_ENV_: string | undefined
 declare const _OT_PD_BUILD_DATE_: string | undefined
+declare const _OT_PD_LATEST_LABWARE_VERSIONS_: Record<string, Number>
 declare const _OT_PD_MIXPANEL_DEV_ID_: string | undefined
 declare const _OT_PD_MIXPANEL_ID_: string | undefined
+declare const _OT_PD_REQUIRED_APP_VERSION_: string | undefined
 declare const _OT_PD_SENTRY_DEV_DSN_: string | undefined
 declare const _OT_PD_SENTRY_DSN_: string | undefined
 declare const _OT_PD_VERSION_: string | undefined