+++ step-generation/src/commandCreators/compound/distribute.ts
@@ -14,7 +14,7 @@ import {
   NONE_LIQUID_CLASS_NAME,
   POSITION_REFERENCE_MAPPED_TO_WELL_ORIGIN,
   SAFE_MOVE_TO_WELL_LOCATION,
-  WATER_LIQUID_CLASS_NAME_V2,
+  WATER_LIQUID_CLASS_NAME,
   WELL_ORIGIN_TOP,
 } from '@opentrons/shared-data'
 
@@ -111,7 +111,6 @@ export const distribute: CommandCreator<DistributeArgs> = (
     blowoutFlowRateUlSec,
     blowoutLocation,
     changeTip,
-    conditioningVolume,
     destLabware,
     destWells,
     dispenseDelay,
@@ -171,6 +170,10 @@ export const distribute: CommandCreator<DistributeArgs> = (
   const disposalVolume =
     args.disposalVolume != null && args.disposalVolume > 0
       ? args.disposalVolume
+      : 0
+  const conditioningVolume =
+    args.conditioningVolume != null && args.conditioningVolume > 0
+      ? args.conditioningVolume
       : 0
   // TODO: Ian 2019-04-19 revisit these pipetteDoesNotExist errors, how to do it DRY?
   if (
@@ -262,7 +265,7 @@ export const distribute: CommandCreator<DistributeArgs> = (
     getAllLiquidClassDefs()
       [
         liquidClass === NONE_LIQUID_CLASS_NAME || liquidClass == null
-          ? WATER_LIQUID_CLASS_NAME_V2
+          ? WATER_LIQUID_CLASS_NAME
           : liquidClass
       ].byPipette?.find(
         ({ pipetteModel }) =>
@@ -572,6 +575,8 @@ export const distribute: CommandCreator<DistributeArgs> = (
     (destWellChunk: string[], chunkIndex: number): CurriedCommandCreator[] => {
       const numDestsPerAsp = destWellChunk.length // can differ on final chunk
       const totalSampleAspirateVolume = volume * numDestsPerAsp
+      const totalGrossAspirateVolume =
+        totalSampleAspirateVolume + disposalVolume + conditioningVolume
       const isFirstChunk = chunkIndex === 0
       const isLastChunk = chunkIndex === destWellChunks.length - 1
       const changeTipNow =
@@ -584,7 +589,7 @@ export const distribute: CommandCreator<DistributeArgs> = (
         ? [
             curryWithoutPython(configureForVolume, {
               pipetteId: pipette,
-              volume: totalSampleAspirateVolume,
+              volume: totalGrossAspirateVolume,
             }),
           ]
         : []
@@ -805,10 +810,7 @@ export const distribute: CommandCreator<DistributeArgs> = (
           liquidClass,
           pipetteSpecs,
           tiprackDefUri: tipRack,
-          targetVolume:
-            totalSampleAspirateVolume +
-            (disposalVolume ?? 0) +
-            (conditioningVolume ?? 0),
+          targetVolume: totalGrossAspirateVolume,
           liquidHandlingAction: 'aspirate',
           byVolumeProperty: 'correctionByVolume',
           defaultValue: 0,
@@ -818,7 +820,7 @@ export const distribute: CommandCreator<DistributeArgs> = (
           liquidClass,
           pipetteSpecs,
           tiprackDefUri: tipRack,
-          targetVolume: conditioningVolume ?? 0,
+          targetVolume: conditioningVolume,
           liquidHandlingAction: 'multiDispense',
           byVolumeProperty: 'correctionByVolume',
           defaultValue: 0,
@@ -839,10 +841,7 @@ export const distribute: CommandCreator<DistributeArgs> = (
       const aspirateCommands = [
         curryWithoutPython(aspirateInPlace, {
           pipetteId: pipette,
-          volume:
-            totalSampleAspirateVolume +
-            (disposalVolume ?? 0) +
-            (conditioningVolume ?? 0),
+          volume: totalGrossAspirateVolume,
           flowRate: aspirateFlowRateUlSec,
           correctionVolume: aspirateCorrectionVolumeForTotalAspiration,
         }),
@@ -885,7 +884,7 @@ export const distribute: CommandCreator<DistributeArgs> = (
           } else if (
             !isFirstWellInChunk &&
             dispenseAirGapVolume > 0 &&
-            (conditioningVolume == null || conditioningVolume === 0)
+            conditioningVolume === 0
           ) {
             airGapInTip = dispenseAirGapVolume
             airGapDispenseFlowRate = dispenseAirGapDispenseFlowRate
@@ -1042,11 +1041,7 @@ export const distribute: CommandCreator<DistributeArgs> = (
           ): CurriedCommandCreator[] =>
             dispenseAirGapVolume > 0 &&
             // don't air gap if not last well in chunk and conditioning volume is present
-            !(
-              wellIndex < destWellChunk.length - 1 &&
-              conditioningVolume != null &&
-              conditioningVolume > 0
-            ) &&
+            !(wellIndex < destWellChunk.length - 1 && conditioningVolume > 0) &&
             // don't air gap if end of full transfer and not changing tip
             !(
               changeTip === 'never' &&